requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
anytree==2.8.0
async-timeout==3.0.1
Babel==2.9.1
boto3==1.18.44
botocore==1.21.44
cachetools==4.2.4
certifi==2022.12.7
chardet==4.0.0
ceda-directory-tree==1.0.2
docutils==0.16
elasticsearch==7.13.1
fsspec==2021.10.0
gcsfs==2021.10.0
idna==2.10
intake-esm==2024.2.6
imagesize==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
multidict==5.1.0
packaging==20.9
pika==1.2.0
pydantic==2.1.1
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.5.0
siphon==0.9
six==1.16.0
snowballstemmer==2.1.0
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.5
